Massive update:  compiles and runs.  Not road tested.
In longer road tests, found that at stop lights, the average speed was only updating once per second.  It
turns out that it was calculating values assuming the timer "tick" was in milliseconds, when it was in
seconds.  The running timer count is the residual milliseconds.  However, it was still producing "valid"
output because it was assuming the GPS reported speed was in meters/sec, but it is actually in
millimeters/sec.  The mistaken factor of 1000 was canceling out.  This has been corrected.

This update massively reorganizes the data tracking.  Structures have been brought in for the active
race and the active leg and all the tracked data has been moved into those structures.  All data values
are manipulated in meters and floating point seconds.  Conversion to miles and mph is done at display
time.  Currently the race and leg structures are hard coded and there is a single leg.  Stats are tracked
at both the race and leg level now.

Initial data structures for race/leg/point definitions have been created, in prep for loading them off
the uSD card.

The start/stop button still starts/stops the timer clocked by the GPS PPS output.  This may eventually be
changed to allow the timer to free run as long as the pps signal is active and change the start/stop
to just timestamp the button press.

On the physical front, the case design has been updated to add a tounge and groove at the joint of the two
halves.  Hopefully this will provide a more positive connection between the two halves.