Updated case with protrusions on the bottom that align with the tabled mount brackets I am using for mounting in the car.  This will hold quite securely.
Code now has the initial framework for syncronizing timing start to a timing mark. Doesn't totally
work yet in that the delay calculations appear to be off.  It will, however delay the start of timing
to near the correct time and the start the raceLeg.  Right now, it immediately stops the leg to capture
the delay time on the LED counter which starts at button press.  Eventually, the LED counter will
start at race leg start.