Begin restructure the display routines to make the movable between displays and on a given display, where applicable. No new display routines added, this just moved the existing ones into functions, seperating them from the OLED panel and the fixed location.
Unfortunately the autoformatter got turned loose on gpc.cpp and I don't care to go undo it, so there are a lot of
whitespace changes.

Added functionality to blink the start button when starting a race leg with a delayed (synced to a start mark)
start.
Fixed bugs in the delayed start functionality.  There was some math that wasn't quite right and, more importantly,
the event timer apparently runs slow.  To get the display to start updating properly at the start mark, the delay
time needed to be shortened by 20%.  Note this is purely a display artifact.  The actual start time is correct.

There is a glitch in the race time display on the LEDs.  Every once in a while, the seconds digit will flash
back to it's prior value.  I am assuming this is an artifact of the timer ticking while the display is updating.
Because the time displayed is based on race start time and not timer start time, there is an offset that needs
to be subtracted.  If the millisecond count rolls over in the middle of display update, and math is done before
the seconds rolls over, time could be wrong for one display update.  Will fix at some point, but its not a
functional problem.