OLED displays are now handled by a list of display items that are called on each display update tick. This will allow the displays to be easily updated by the (soon to be written) state machine.  The U8G2 objects now use dynamic buffers, which means each display has its own independent buffer now. Memory usage obviously goes up, but its needed so that we don't have worry about the order we draw displayObjects -- which screen they go to.  We just clear all the screen buffers, draw all the displayObjects, the update all the screens.
Added large displays for the various race leg values.  Will need to update naming to note they are
the leg values when the race value display functions are added.  Added display on the LED for all
the leg values as well.

GPS config now ensures the dynamics model is automotive, and we disable the jamming filter because
we don't care.

Started adding code to load races for uSD card, but for that to be useful, we need the menu system
in place, which needs the state machine to be written, which needed the display routines to be moved
to displayObject lists.