Basic flesh out of the state machine code.  The start/stop button lighting is fully controlled by the state machine now. The button lighting is controlled by the state machine, except for the light when press which is in the keypad logic. The keypad buttons are now locked out during an active leg.  The start/stop button will only start a leg when in race start mode.  When not in race start mode, it operates as a normal button for menuing use.
Basic state machine flow is in place.  Next step will be to start building out the displays for each of the states and adding
state transition logic for some of the states that currently just pass through.

The state machine has three main blocks:

* automatic state transition
* key based state transition
* hardware control

The automatic state transition block handles state transitions caused by async state change elsewhere in the code.  The
transition from delayed start to leg in progress is a good example.

The key based transition block handles transitions driven by key presses.  Note that pressing the start/stop button to
start a race leg doesn't not trigger a transition due to the key press.  It triggers a transition in the automatic
transition block due to the race timing code starting and the race leg getting started.

The hardware control block is what updates the lighting on the buttons.  It is a rather complex and finicky piece of code,
in order to handle all the cases for the start/stop button.

With this change, the code now has two indepently operating chunks of code.  The race timing code and display update code
are completely event driven.  The state machine code just runs in the arduino loop() function.  The state machine is not
time sensitive.  All time sensitive code is handled by the even timer and the most sensitive -- race timing -- is
handled by the GPS timestamps.