Started laying out leg and race summary and detail screens.
I think I fixed a bug where sometimes the first push of the start/stop button wouldn't
register.  Quite a problem if you are trying to start a race and it doesn't register.  I
believe it was a race between reading the signal and the interrupt triggering.  Seperated
out flags for keypad press and start/stop press.

Just tested and and the start/stop bug is still there...  <grrr>

Race average speed on the second leg of a race is completely wrong.  Some race values
are either getting reset, or not propagated from leg to race at the end of a leg.