Fixed bug where odometer offset at the start of a leg was being stored to the top level race instead of the active leg.  This would have resulted in very erroneous timing in situations where the computer had been running for a long period before race start.
Disabled the auto-end when the leg distance matched distance driven.  This will rarely be the case.
Will probably make it a config option later.

Timestamp calculations in updateRace() were completely misbehaving.  The seconds field was getting
overwritten somehow -- I think this may be a compiler bug.  Switched to converting timestampts to
fixed point interger representations, which works well.  Will probably do that anywhere timestamp
math is done.

Added a function to dump all the data in a race structure to the serial port, for debugging.

Added saving of a race checkpoint file.  Currently saved at completion of a leg.  Will allow the
computer to be powered down between legs and still keep the overall race stats.  Started the
work on the checkpoint load.