Switched the keypad colors from current leg speed delta to race speed delta.  Should be configurable in the future, but for now, the race delta is what is important.
System can now load race checkpoints on power up and restore the race state between
legs.

In bench testing I now see the display occasionally lag.  I don't know whether this is due
to lack of GPS updates since the display contents are all driven by GPS data, or if the event
timer is somehow lagging, or if something is getting hung up somewhere.  It seemed to start,
or at least become prevalent when saving and loading of race checkpoints was added.  Since
the SD card is the only thing that shares is MISO line with the GPS, it makes me wonder if
there is some sort of contention going on.