Corrected a bug the meant the displayed race leg and selected race leg were out of sync so the race would not end when the leg completed after a checkpoint load.
Confirmed that the GPS data update jerkiness is not due to the event timer getting hung
up.  Race time on the LED display progresses smoothly -- it is not dependent on GPS data.
It also doesn't us the MISO signal.  So, its not a problem writing on the SPI bus.  Still
could be a read problem.  May move the SD card to SPI2, though that will take some serious
bodge work.