Started adding framework for configurable disable panels.
if there is a problem reading the SD card, at any point during loading of race definitions,
it will fall back to an internal set of pre-defined races.  This won't be useful for others,
but works as a fallback in case of catastrophic SD card issues for the upcoming BBORR.

When accessing the SD card, we now disable all the other access to the SPI bus.  The problem
was that the reading/writing of the SD card was in the async program loop, while the other
accesses were event driven.  You could have the event callback interrupt trigger in the
middle of and SD card read or write and a display update or, especially, a GPS read could
interfere and trash both transactions.  We don't access the SD card during timing critical
activities, so it doesn't cause any problems to freeze the displays for a few 10s of milli
seconds.  If we add race logging in the future, log updates would be event driven, which
would prevent them from colliding.